feat: add submission indicator and task status chip components
- Implemented SubmissionIndicator component to show submission status (idle, sending, success, error).
- Created TaskStatusChip component to display task execution status (not executed, correct, incorrect).
- Developed TaskSummaryCard component to summarize task details, including submission status and task status chip.
- Introduced TaskView component to manage code and task details layout.
- Refactored page structure to utilize new components for better organization and readability.
- Enhanced FinishedScreen to display task results and submission status.
