+++ packages/vulpes/public/icons/XIcon.tsx
@@ -1,0 +1,19 @@
+import { exportPages } from "next/dist/build/worker";
+
+const XIcon = ({ className }: { className?: string }) => (
+  <svg
+    className={className}
+    fill="none"
+    viewBox="0 0 24 24"
+    stroke="currentColor"
+  >
+    <path
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={2}
+      d="M6 18L18 6M6 6l12 12"
+    />
+  </svg>
+);
+
+export default XIcon;
