+++ packages/vulpes/src/app/(list)/list/[ID]/components/DetailsSection.tsx
@@ -1,0 +1,35 @@
+import { ITask } from "@/@types/Task";
+import { Box } from "@mui/material";
+import MDEditor from "@uiw/react-md-editor";
+
+interface IDetailsSection {
+  task: ITask | null;
+}
+
+export default function DetailsSection({ task }: IDetailsSection) {
+  return (
+    <Box
+      sx={{
+        display: "flex",
+        flexDirection: "column",
+        height: "100%",
+        minHeight: 0,
+        borderRadius: "8px",
+        overflow: "hidden",
+        backgroundColor: "transparent",
+        color: "white",
+      }}
+    >
+      <Box sx={{ flex: 1, minHeight: 0, overflowY: "auto" }}>
+        <MDEditor.Markdown
+          source={task?.description || ""}
+          style={{
+            padding: "8px 16px",
+            backgroundColor: "transparent",
+            color: "white",
+          }}
+        />
+      </Box>
+    </Box>
+  );
+}
