+++ packages/vulpes/src/app/(list)/list/[ID]/page.tsx
@@ -42,38 +42,13 @@ import ErrorIcon from "@mui/icons-material/Error";
 import SendIcon from "@mui/icons-material/Send";
 import { AnimatePresence, motion } from "framer-motion";
 import { COLORS } from "@/utils/colors";
+import CodeSection from "./components/CodeSection";
+import DetailsSection from "./components/DetailsSection";
+import TaskView from "./components/TaskView";
+import TaskSummaryCard from "./components/TaskSummaryCard";
+import FinishedScreen from "./components/FinishedScreen";
 
-const CheckIcon = ({ className }: { className?: string }) => (
-  <svg
-    className={className}
-    fill="none"
-    viewBox="0 0 24 24"
-    stroke="currentColor"
-  >
-    <path
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth={2}
-      d="M5 13l4 4L19 7"
-    />
-  </svg>
-);
-
-const XIcon = ({ className }: { className?: string }) => (
-  <svg
-    className={className}
-    fill="none"
-    viewBox="0 0 24 24"
-    stroke="currentColor"
-  >
-    <path
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth={2}
-      d="M6 18L18 6M6 6l12 12"
-    />
-  </svg>
-);
+export type TaskSubmissionStatus = "idle" | "sending" | "success" | "error";
 
 export default function Page() {
   return (
@@ -104,820 +79,6 @@ function LayoutBox({ children }: { children: React.ReactNode }) {
       }}
     >
       {children}
-    </Box>
-  );
-}
-
-interface ICollapsibleHeaderProps {
-  icon: React.ReactNode;
-  title: string;
-  collapsed: boolean;
-  onToggle: () => void;
-  trailing?: React.ReactNode;
-}
-
-function CollapsibleHeader({
-  icon,
-  title,
-  collapsed,
-  onToggle,
-  trailing,
-}: ICollapsibleHeaderProps) {
-  return (
-    <Box
-      onClick={onToggle}
-      sx={{
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "space-between",
-        px: 2,
-        py: 1,
-        bgcolor: "#252526",
-        borderBottom: collapsed ? "none" : "1px solid",
-        borderColor: "rgba(255,255,255,0.08)",
-        cursor: "pointer",
-        userSelect: "none",
-        borderRadius: collapsed ? "8px" : "8px 8px 0 0",
-        transition: "background-color 0.15s ease",
-        "&:hover": { bgcolor: "#2d2d30" },
-      }}
-    >
-      <Box sx={{ display: "flex", alignItems: "center", gap: 1 }}>
-        <Box sx={{ color: COLORS.dark.primary[500], display: "flex" }}>
-          {icon}
-        </Box>
-        <Typography
-          variant="body2"
-          sx={{ color: "#d4d4d4", fontWeight: 600, letterSpacing: 0.2 }}
-        >
-          {title}
-        </Typography>
-      </Box>
-      <Box sx={{ display: "flex", alignItems: "center", gap: 1 }}>
-        {trailing}
-        <IconButton
-          size="small"
-          sx={{ color: "#9e9e9e", p: 0.25 }}
-          onClick={(e) => {
-            e.stopPropagation();
-            onToggle();
-          }}
-        >
-          {collapsed ? (
-            <ExpandMoreIcon fontSize="small" />
-          ) : (
-            <ExpandLessIcon fontSize="small" />
-          )}
-        </IconButton>
-      </Box>
-    </Box>
-  );
-}
-
-interface ICodeSection {
-  code: string;
-  setCode: (code: string) => void;
-  submissionStatus: "success" | "error" | null;
-  lastResults: ITestCaseResult[];
-  compileErrors: ICompileError[];
-  isRunning: boolean;
-}
-
-function CodeSection({
-  code,
-  setCode,
-  submissionStatus,
-  lastResults,
-  compileErrors,
-  isRunning,
-}: ICodeSection) {
-  const [editorCollapsed, setEditorCollapsed] = useState(false);
-  const [outputCollapsed, setOutputCollapsed] = useState(false);
-
-  function handleEditorDidMount(_editor: any, monacoInstance: any) {
-    registerPortugolLanguage(monacoInstance);
-    monacoInstance.editor.setTheme("vs-dark");
-  }
-
-  const statusBadge =
-    submissionStatus === "success" ? (
-      <span className="flex items-center text-xs font-bold text-green-500 bg-green-500/10 px-2 py-0.5 rounded-full border border-green-500/20">
-        <CheckIcon className="w-3.5 h-3.5 mr-1" /> SUCESSO
-      </span>
-    ) : submissionStatus === "error" ? (
-      <span className="flex items-center text-xs font-bold text-red-500 bg-red-500/10 px-2 py-0.5 rounded-full border border-red-500/20">
-        <XIcon className="w-3.5 h-3.5 mr-1" /> FALHA
-      </span>
-    ) : null;
-
-  return (
-    <Box
-      sx={{
-        display: "flex",
-        flexDirection: "column",
-        minHeight: 0,
-        height: "100%",
-        gap: 1,
-      }}
-    >
-      <Box
-        sx={{
-          display: "flex",
-          flexDirection: "column",
-          flex: editorCollapsed ? "0 0 auto" : "1 1 70%",
-          minHeight: 0,
-          borderRadius: "8px",
-          overflow: "hidden",
-          bgcolor: "#1e1e1e",
-          border: "1px solid rgba(255,255,255,0.05)",
-        }}
-      >
-        <CollapsibleHeader
-          icon={<CodeIcon fontSize="small" />}
-          title="Editor"
-          collapsed={editorCollapsed}
-          onToggle={() => setEditorCollapsed((v) => !v)}
-        />
-        {!editorCollapsed && (
-          <Box sx={{ flex: 1, minHeight: 240 }}>
-            <Editor
-              height="100%"
-              theme="vs-dark"
-              language="portugol"
-              value={code}
-              onChange={(value) => setCode(value || "")}
-              onMount={handleEditorDidMount}
-              options={{
-                fontSize: 14,
-                tabSize: 2,
-                wordWrap: "on",
-                minimap: { enabled: true },
-                lineNumbers: "on",
-                renderWhitespace: "selection",
-                automaticLayout: true,
-                tabCompletion: "on",
-                cursorStyle: "line",
-                scrollBeyondLastLine: false,
-              }}
-            />
-          </Box>
-        )}
-      </Box>
-
-      <Box
-        sx={{
-          display: "flex",
-          flexDirection: "column",
-          flex: outputCollapsed ? "0 0 auto" : "1 1 30%",
-          minHeight: 0,
-          borderRadius: "8px",
-          overflow: "hidden",
-          bgcolor: "#1e1e1e",
-          border: "1px solid rgba(255,255,255,0.05)",
-        }}
-      >
-        <CollapsibleHeader
-          icon={<TerminalIcon fontSize="small" />}
-          title="Terminal / Resultados"
-          collapsed={outputCollapsed}
-          onToggle={() => setOutputCollapsed((v) => !v)}
-          trailing={statusBadge}
-        />
-        {!outputCollapsed && (
-          <div className="flex-1 overflow-y-auto p-4 space-y-2 bg-[#1e1e1e] min-h-[160px]">
-            {compileErrors.length === 0 &&
-              lastResults.length === 0 &&
-              !isRunning && (
-                <div className="flex h-full items-center justify-center text-gray-500 text-sm">
-                  Execute o código para ver os resultados.
-                </div>
-              )}
-            {isRunning && (
-              <div className="flex h-full items-center justify-center text-gray-400 text-sm animate-pulse">
-                Executando testes...
-              </div>
-            )}
-
-            {!isRunning && compileErrors.length > 0 && (
-              <div className="bg-[#1e1e1e] rounded border border-red-500/30 overflow-hidden">
-                <div className="flex items-center px-3 py-2 bg-red-500/10 border-b border-red-500/20">
-                  <XIcon className="w-4 h-4 text-red-500 mr-2" />
-                  <span className="text-sm font-semibold text-red-400">
-                    {compileErrors.length === 1
-                      ? "1 erro de compilação"
-                      : `${compileErrors.length} erros de compilação`}
-                  </span>
-                </div>
-                <ul className="px-4 py-2 space-y-1 text-xs font-mono text-red-300">
-                  {compileErrors.map((err, idx) => (
-                    <li key={idx} className="leading-relaxed">
-                      <span className="text-gray-500 mr-2">
-                        [{err.kind === "parse" ? "sintaxe" : "semântico"}]
-                      </span>
-                      <span className="text-gray-400 mr-2">
-                        linha {err.line}, coluna {err.column}:
-                      </span>
-                      <span>{err.message}</span>
-                    </li>
-                  ))}
-                </ul>
-                <div className="px-4 py-2 text-[11px] text-gray-500 border-t border-gray-700">
-                  Corrija os erros acima para que os testes sejam executados.
-                </div>
-              </div>
-            )}
-
-            {!isRunning &&
-              compileErrors.length === 0 &&
-              lastResults.map((result, index) => (
-                <div
-                  key={index}
-                  className="flex flex-col bg-[#2d2d2d] rounded bg-opacity-40 overflow-hidden"
-                >
-                  <div
-                    className={`flex items-center px-3 py-2 border-l-4 ${
-                      result.passed
-                        ? "border-green-500 bg-green-500/5"
-                        : "border-red-500 bg-red-500/5"
-                    }`}
-                  >
-                    <span className="mr-3">
-                      {result.passed ? (
-                        <CheckIcon className="w-5 h-5 text-green-500" />
-                      ) : (
-                        <XIcon className="w-5 h-5 text-red-500" />
-                      )}
-                    </span>
-                    <span
-                      className={`text-sm font-medium ${
-                        result.passed ? "text-green-400" : "text-red-400"
-                      }`}
-                    >
-                      Teste {index + 1}
-                    </span>
-                    <span className="ml-auto text-xs text-gray-500">
-                      {result.passed ? "Passou" : "Falhou"}
-                    </span>
-                  </div>
-
-                  {!result.passed && (
-                    <div className="px-4 py-2 bg-[#1e1e1e] bg-opacity-50 text-xs font-mono border-t border-gray-700 text-gray-300">
-                      <div className="grid grid-cols-2 gap-4">
-                        <div>
-                          <span className="block text-gray-500 mb-0.5">
-                            Esperado:
-                          </span>
-                          <div className="bg-gray-800 p-1 rounded text-green-300">
-                            {result.expectedOutput}
-                          </div>
-                        </div>
-                        <div>
-                          <span className="block text-gray-500 mb-0.5">
-                            Obtido:
-                          </span>
-                          <div className="bg-gray-800 p-1 rounded text-red-300">
-                            {result.actualOutput}
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  )}
-                </div>
-              ))}
-          </div>
-        )}
-      </Box>
-    </Box>
-  );
-}
-
-interface IDetailsSection {
-  task: ITask | null;
-}
-
-function DetailsSection({ task }: IDetailsSection) {
-  return (
-    <Box
-      sx={{
-        display: "flex",
-        flexDirection: "column",
-        height: "100%",
-        minHeight: 0,
-        borderRadius: "8px",
-        overflow: "hidden",
-        backgroundColor: "transparent",
-        color: "white",
-      }}
-    >
-      <Box sx={{ flex: 1, minHeight: 0, overflowY: "auto" }}>
-        <MDEditor.Markdown
-          source={task?.description || ""}
-          style={{
-            padding: "8px 16px",
-            backgroundColor: "transparent",
-            color: "white",
-          }}
-        />
-      </Box>
-    </Box>
-  );
-}
-
-interface ITaskViewProps {
-  task: ITask;
-  code: string;
-  setCode: (code: string) => void;
-  submissionStatus: "success" | "error" | null;
-  lastResults: ITestCaseResult[];
-  compileErrors: ICompileError[];
-  isRunning: boolean;
-}
-
-function TaskView({
-  task,
-  code,
-  setCode,
-  submissionStatus,
-  lastResults,
-  compileErrors,
-  isRunning,
-}: ITaskViewProps) {
-  return (
-    <Box
-      sx={{
-        width: "100%",
-        height: "100%",
-        display: "grid",
-        gap: 2,
-        "@media (min-width: 1281px)": {
-          gridTemplateColumns: "minmax(0, 3fr) minmax(0, 2fr)",
-          gridTemplateRows: "1fr",
-        },
-        "@media (max-width: 1280px)": {
-          gridTemplateColumns: "1fr",
-          gridTemplateRows: "auto 1fr",
-        },
-      }}
-    >
-      <CodeSection
-        code={code}
-        setCode={setCode}
-        submissionStatus={submissionStatus}
-        lastResults={lastResults}
-        compileErrors={compileErrors}
-        isRunning={isRunning}
-      />
-      <DetailsSection task={task} />
-    </Box>
-  );
-}
-
-type TaskSubmissionStatus = "idle" | "sending" | "success" | "error";
-
-interface ITaskSummaryCardProps {
-  task: ITask;
-  index: number;
-  code: string;
-  hasResults: boolean;
-  isCorrect: boolean;
-  submissionStatus: TaskSubmissionStatus;
-  onReview: () => void;
-}
-
-function TaskStatusChip({
-  hasResults,
-  isCorrect,
-}: {
-  hasResults: boolean;
-  isCorrect: boolean;
-}) {
-  if (!hasResults) {
-    return (
-      <Chip
-        label="Não executada"
-        size="small"
-        sx={{
-          bgcolor: "rgba(255,255,255,0.06)",
-          color: "#cfd8dc",
-          border: "1px solid rgba(255,255,255,0.12)",
-          fontWeight: 600,
-        }}
-      />
-    );
-  }
-
-  if (isCorrect) {
-    return (
-      <Chip
-        icon={<CheckCircleIcon sx={{ color: "#66bb6a !important" }} />}
-        label="Correta"
-        size="small"
-        sx={{
-          bgcolor: "rgba(102,187,106,0.12)",
-          color: "#a5d6a7",
-          border: "1px solid rgba(102,187,106,0.25)",
-          fontWeight: 600,
-        }}
-      />
-    );
-  }
-
-  return (
-    <Chip
-      icon={<ErrorIcon sx={{ color: "#ef5350 !important" }} />}
-      label="Incorreta"
-      size="small"
-      sx={{
-        bgcolor: "rgba(239,83,80,0.12)",
-        color: "#ef9a9a",
-        border: "1px solid rgba(239,83,80,0.25)",
-        fontWeight: 600,
-      }}
-    />
-  );
-}
-
-function SubmissionIndicator({ status }: { status: TaskSubmissionStatus }) {
-  if (status === "idle") return null;
-
-  if (status === "sending") {
-    return (
-      <Box sx={{ display: "flex", alignItems: "center", gap: 0.75 }}>
-        <CircularProgress size={14} sx={{ color: COLORS.dark.primary[500] }} />
-        <Typography variant="caption" sx={{ color: "#cfd8dc" }}>
-          Enviando...
-        </Typography>
-      </Box>
-    );
-  }
-
-  if (status === "success") {
-    return (
-      <Box sx={{ display: "flex", alignItems: "center", gap: 0.5 }}>
-        <CheckCircleIcon sx={{ color: "#66bb6a", fontSize: 18 }} />
-        <Typography variant="caption" sx={{ color: "#a5d6a7" }}>
-          Enviada
-        </Typography>
-      </Box>
-    );
-  }
-
-  return (
-    <Box sx={{ display: "flex", alignItems: "center", gap: 0.5 }}>
-      <ErrorIcon sx={{ color: "#ef5350", fontSize: 18 }} />
-      <Typography variant="caption" sx={{ color: "#ef9a9a" }}>
-        Falhou
-      </Typography>
-    </Box>
-  );
-}
-
-function TaskSummaryCard({
-  task,
-  index,
-  code,
-  hasResults,
-  isCorrect,
-  submissionStatus,
-  onReview,
-}: ITaskSummaryCardProps) {
-  const [codeCollapsed, setCodeCollapsed] = useState(true);
-  const lines = code.split("\n").length;
-
-  return (
-    <Card
-      sx={{
-        bgcolor: "#1e272c",
-        border: "1px solid rgba(255,255,255,0.06)",
-        borderRadius: 2,
-        color: "#fff",
-      }}
-    >
-      <CardContent sx={{ p: 0, "&:last-child": { pb: 0 } }}>
-        <Box
-          sx={{
-            display: "flex",
-            alignItems: "center",
-            gap: 2,
-            px: 2.5,
-            py: 2,
-            borderBottom: codeCollapsed
-              ? "none"
-              : "1px solid rgba(255,255,255,0.06)",
-          }}
-        >
-          <Box
-            sx={{
-              width: 40,
-              height: 40,
-              borderRadius: "50%",
-              bgcolor: COLORS.dark.primary[500],
-              color: "#fff",
-              display: "flex",
-              alignItems: "center",
-              justifyContent: "center",
-              fontWeight: 700,
-              flexShrink: 0,
-            }}
-          >
-            {index + 1}
-          </Box>
-          <Box sx={{ flex: 1, minWidth: 0 }}>
-            <Typography
-              variant="subtitle1"
-              sx={{ fontWeight: 600, color: "#fff", lineHeight: 1.2 }}
-              noWrap
-            >
-              {task.title}
-            </Typography>
-            <Typography
-              variant="caption"
-              sx={{ color: "#90a4ae", display: "block" }}
-            >
-              Tarefa {index + 1} · {lines} linha{lines === 1 ? "" : "s"} de
-              código
-            </Typography>
-          </Box>
-          <SubmissionIndicator status={submissionStatus} />
-          <TaskStatusChip hasResults={hasResults} isCorrect={isCorrect} />
-          <Tooltip title="Revisar tarefa">
-            <IconButton
-              size="small"
-              onClick={onReview}
-              sx={{ color: "#90a4ae" }}
-            >
-              <CodeIcon fontSize="small" />
-            </IconButton>
-          </Tooltip>
-          <IconButton
-            size="small"
-            onClick={() => setCodeCollapsed((v) => !v)}
-            sx={{ color: "#90a4ae" }}
-          >
-            {codeCollapsed ? (
-              <ExpandMoreIcon fontSize="small" />
-            ) : (
-              <ExpandLessIcon fontSize="small" />
-            )}
-          </IconButton>
-        </Box>
-
-        <Collapse in={!codeCollapsed} unmountOnExit>
-          <Box sx={{ bgcolor: "#1e1e1e", height: 260 }}>
-            <Editor
-              height="100%"
-              theme="vs-dark"
-              language="portugol"
-              value={code}
-              onMount={(_e, monacoInstance) => {
-                registerPortugolLanguage(monacoInstance);
-                monacoInstance.editor.setTheme("vs-dark");
-              }}
-              options={{
-                readOnly: true,
-                fontSize: 13,
-                tabSize: 2,
-                wordWrap: "on",
-                minimap: { enabled: false },
-                lineNumbers: "on",
-                automaticLayout: true,
-                scrollBeyondLastLine: false,
-              }}
-            />
-          </Box>
-        </Collapse>
-      </CardContent>
-    </Card>
-  );
-}
-
-interface IFinishedScreenProps {
-  tasks: ITask[];
-  codesByTaskId: Record<string, string>;
-  resultsByTaskId: Record<string, ITestCaseResult[]>;
-  submissionStatusByTaskId: Record<string, TaskSubmissionStatus>;
-  isSubmittingList: boolean;
-  canRegisterSubmission: boolean;
-  hasSubmittedList: boolean;
-  onReviewTask: (index: number) => void;
-  onSubmitList: () => void;
-}
-
-function FinishedScreen({
-  tasks,
-  codesByTaskId,
-  resultsByTaskId,
-  submissionStatusByTaskId,
-  isSubmittingList,
-  hasSubmittedList,
-  onReviewTask,
-  onSubmitList,
-  canRegisterSubmission,
-}: IFinishedScreenProps) {
-  const totalLines = tasks.reduce(
-    (acc, t) => acc + (codesByTaskId[t.taskId]?.split("\n").length ?? 0),
-    0,
-  );
-
-  const correctCount = tasks.reduce((acc, t) => {
-    const results = resultsByTaskId[t.taskId];
-    if (!results || results.length === 0) return acc;
-    return acc + (results.every((r) => r.passed) ? 1 : 0);
-  }, 0);
-
-  const sentCount = tasks.reduce(
-    (acc, t) =>
-      acc + (submissionStatusByTaskId[t.taskId] === "success" ? 1 : 0),
-    0,
-  );
-  const failedCount = tasks.reduce(
-    (acc, t) => acc + (submissionStatusByTaskId[t.taskId] === "error" ? 1 : 0),
-    0,
-  );
-
-  return (
-    <Box
-      sx={{
-        width: "100%",
-        height: "100%",
-        overflowY: "auto",
-        px: { xs: 2, md: 4 },
-        py: 4,
-      }}
-    >
-      <Box sx={{ maxWidth: 920, mx: "auto" }}>
-        <Box
-          sx={{
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "space-between",
-            gap: 2,
-            mb: 1,
-            flexWrap: "wrap",
-          }}
-        >
-          <Box
-            sx={{
-              display: "flex",
-              alignItems: "center",
-              gap: 2,
-              color: COLORS.dark.primary[500],
-            }}
-          >
-            <CheckCircleIcon sx={{ fontSize: 40 }} />
-            <Typography variant="h4" sx={{ fontWeight: 700, color: "#fff" }}>
-              Lista finalizada!
-            </Typography>
-          </Box>
-
-          <Button
-            variant="contained"
-            size="large"
-            onClick={onSubmitList}
-            disabled={isSubmittingList || tasks.length === 0}
-            startIcon={
-              isSubmittingList ? (
-                <CircularProgress size={18} sx={{ color: "#fff" }} />
-              ) : (
-                <SendIcon />
-              )
-            }
-            sx={{
-              bgcolor: COLORS.dark.primary[500],
-              "&:hover": { bgcolor: COLORS.dark.primary[600] },
-              fontWeight: 600,
-              textTransform: "none",
-              px: 3,
-            }}
-          >
-            {isSubmittingList
-              ? "Enviando lista..."
-              : hasSubmittedList
-                ? "Reenviar lista"
-                : "Enviar lista"}
-          </Button>
-        </Box>
-
-        <Typography variant="body1" sx={{ color: "#cfd8dc", mb: 3 }}>
-          Você concluiu todas as tarefas desta lista. Revise suas soluções
-          abaixo e clique em <b>Enviar lista</b> para submeter cada tarefa
-          individualmente.
-        </Typography>
-
-        <Box
-          sx={{
-            display: "grid",
-            gap: 2,
-            gridTemplateColumns: {
-              xs: "1fr",
-              sm: "repeat(2, 1fr)",
-              md: "repeat(4, 1fr)",
-            },
-            mb: 3,
-          }}
-        >
-          <Card sx={{ bgcolor: "#1e272c", color: "#fff", borderRadius: 2 }}>
-            <CardContent>
-              <Typography variant="caption" sx={{ color: "#90a4ae" }}>
-                Tarefas
-              </Typography>
-              <Typography variant="h4" sx={{ fontWeight: 700 }}>
-                {tasks.length}
-              </Typography>
-            </CardContent>
-          </Card>
-          <Card sx={{ bgcolor: "#1e272c", color: "#fff", borderRadius: 2 }}>
-            <CardContent>
-              <Typography variant="caption" sx={{ color: "#90a4ae" }}>
-                Corretas
-              </Typography>
-              <Typography
-                variant="h4"
-                sx={{ fontWeight: 700, color: "#a5d6a7" }}
-              >
-                {correctCount}
-                <Typography
-                  component="span"
-                  variant="body2"
-                  sx={{ color: "#90a4ae", ml: 0.5 }}
-                >
-                  / {tasks.length}
-                </Typography>
-              </Typography>
-            </CardContent>
-          </Card>
-          <Card sx={{ bgcolor: "#1e272c", color: "#fff", borderRadius: 2 }}>
-            <CardContent>
-              <Typography variant="caption" sx={{ color: "#90a4ae" }}>
-                Linhas de código
-              </Typography>
-              <Typography variant="h4" sx={{ fontWeight: 700 }}>
-                {totalLines}
-              </Typography>
-            </CardContent>
-          </Card>
-          <Card sx={{ bgcolor: "#1e272c", color: "#fff", borderRadius: 2 }}>
-            <CardContent>
-              <Typography variant="caption" sx={{ color: "#90a4ae" }}>
-                Enviadas
-              </Typography>
-              <Typography
-                variant="h4"
-                sx={{
-                  fontWeight: 700,
-                  color: failedCount > 0 ? "#ef9a9a" : "#a5d6a7",
-                }}
-              >
-                {sentCount}
-                <Typography
-                  component="span"
-                  variant="body2"
-                  sx={{ color: "#90a4ae", ml: 0.5 }}
-                >
-                  / {tasks.length}
-                </Typography>
-              </Typography>
-              {failedCount > 0 && (
-                <Typography
-                  variant="caption"
-                  sx={{ color: "#ef9a9a", display: "block" }}
-                >
-                  {failedCount} falha{failedCount === 1 ? "" : "s"}
-                </Typography>
-              )}
-            </CardContent>
-          </Card>
-        </Box>
-
-        <Typography
-          variant="h6"
-          sx={{ color: "#fff", fontWeight: 600, mb: 1.5 }}
-        >
-          Suas soluções
-        </Typography>
-        <Box sx={{ display: "flex", flexDirection: "column", gap: 1.5 }}>
-          {tasks.map((task, index) => {
-            const results = resultsByTaskId[task.taskId] ?? [];
-            const hasResults = results.length > 0;
-            const isCorrect = hasResults && results.every((r) => r.passed);
-
-            return (
-              <TaskSummaryCard
-                key={task.taskId}
-                task={task}
-                index={index}
-                code={codesByTaskId[task.taskId] ?? ""}
-                hasResults={hasResults}
-                isCorrect={isCorrect}
-                submissionStatus={
-                  submissionStatusByTaskId[task.taskId] ?? "idle"
-                }
-                onReview={() => onReviewTask(index)}
-              />
-            );
-          })}
-        </Box>
-      </Box>
     </Box>
   );
 }
@@ -964,6 +125,10 @@ function ListRunner() {
   const setCurrentCode = (newCode: string) => {
     if (!currentTask) return;
     setCodesByTaskId((prev) => ({ ...prev, [currentTask.taskId]: newCode }));
+
+    // Invalida os resultados do teste se o código foi alterado
+    setResultsByTaskId((prev) => ({ ...prev, [currentTask.taskId]: [] }));
+    setSubmissionStatus(null);
   };
 
   const blockSubmissions = () => {
@@ -975,7 +140,14 @@ function ListRunner() {
 
   const checkIfListIsValid = (pList: IList) => {
     const submissionLimit = pList.submissionLimit;
-    const currentSubmissions = pList.submissions ?? [];
+    // currentSubmissions deve ser o agrupamento de submissões feitas com diferentes submittedAt
+    const submissionsTime = pList.submissions.map(
+      (submission) => new Date(submission.submittedAt),
+    );
+    const uniqueSubmissionTimes = new Set(
+      submissionsTime.map((date) => date.getTime()),
+    );
+    const currentSubmissions = Array.from(uniqueSubmissionTimes);
 
     if (submissionLimit) {
       return currentSubmissions.length < submissionLimit;
@@ -986,10 +158,17 @@ function ListRunner() {
 
   const getList = async (pListId: string) => {
     try {
+      // Se o usuário for um estudante, as submissões buscadas serão somente as que pertencem a ele
       const response = await API.get(`/list/${pListId}`);
       const responseList: IList = response.data;
 
-      if (!checkIfListIsValid(responseList)) blockSubmissions();
+      const isListValid = checkIfListIsValid(responseList);
+
+      if (!isListValid) {
+        blockSubmissions();
+      } else {
+        setCanRegisterSubmission(true);
+      }
 
       setList(responseList);
     } catch (e) {
@@ -1130,7 +309,13 @@ function ListRunner() {
   };
 
   const handleSubmitList = async () => {
-    if (isSubmittingList || !listId || tasksInList.length === 0) return;
+    if (
+      isSubmittingList ||
+      !listId ||
+      tasksInList.length === 0 ||
+      !canRegisterSubmission
+    )
+      return;
 
     setIsSubmittingList(true);
 
@@ -1178,6 +363,8 @@ function ListRunner() {
         }));
       }
     }
+
+    await getList(listId as string);
 
     setIsSubmittingList(false);
     setHasSubmittedList(true);
